traffic-monitoring/requirements.txt
ultralytics>=8.0.0
numpy>=1.22.2
opencv-python>=4.5.4.60
torch>=1.13.0
torchvision>=0.14.0
deep-sort-realtime>=1.3.2
pyyaml>=6.0